Set-up for testing, added tagdoc methods and tests for encounters
Added a number of directories and supporting files so that tests could
be run with pytest.  Added methods to Encounter objects to parse tagdoc
and report tag type, position, orgin and colour.  Also moved field
lookup from models.py so constants.py so that they could shared across
the application.