Added angler recoveries to tags applied in project view
Added tag recoveries from the general public to the view that summarizes
tags applied in a particular project.  Modified the template to include
a tab for recoveries from the general public and re-factored the view to
be simpler by creating several helper functions that call raw sql.
Tests need to be written to confirm that these functions work as
expected.