Added tests to verify raw sql functions for tags applied in a project
This commit adds two tests that were written to verify the raw sql
queries written to retrieve tags applied in a specific project and
subsequently recovered by the OMNR or the general public.  This commit
also includes another test written to verify that angler recoveries
without lat-lon are not included in points dictionary - can't plot
points that don't exist.