Update tag detail and tag contains templates to include angler returns in N
The number of observations originally reported did not include angler
recaps.  Update the get_tagdetails() function and views for tag details
and tag_contains to include the variable nobs.  Added integration tests
to verify that N={nobs} is included in the rendered templates and
accurately reflects the number of tags presented.