Added Create Person button to angler list and associated  tests
If no anglers can be found matching the filter criteria - we need a way
to add new anglers. A button to a (yet to be created form) was included
with the message that no people could be found.  Wrote several
integration tests to verify that angler filters are working as expected
and that the button is available, and only available when no people are
returned by the filters.