Added Follow-up required button to Angler Report List
A button was added to angler report list that will appear on reports that
have an outstanding follow-up (this functionality has always existed on
the report list, but not the angler-reports detail page).  Tests were
added to verify that the elements on the angler reports detail page
render as expected and that the "Follow-up Required" button appears when,
and only when its needed.