Added angler form, views and templates
Added a form to create new or edit existing anglers.  The form is very
basic and no tests have been implemented to verify functionality.  Need
logic to check for anglers with same name and prompt for appropriate
course of action - update the existing angler or create another with the
same name.