Finalized forms for anglers including tests
Added a second form for anglers that inherits from the first and is used
to create new users - including logic allowing users with the same first
and last name to be saved as different people.  This commit also
includes tests of the new forms for anglers.  Small changes to settings
file to activate django debug-toolbar and render add-user button if the
angler list is filtered (not just if no anglers are returned).