Added tests for report detail page.
Add a number of tests to verify that all of the required and conditional
elements associated with tag report render appropriately.  4 tests
currently fail as they are associated with report elements that have not
been included in report model (DCR, fishing effort and associated file)
or urls that need to be updated (edit_report).