Added html file with lat-long example
The example file contains a form that can be used to capture tag
recovery information in a number of different formats, including a
clickable map, and lat-long in decimal degrees, degrees, decimal
minutes, and degree-minutes-seconds.  Not tested or even integrated into
tfat yet, but it does seem to work as expected.

Additionally, there is no validation or verification to ensure data
standards at this point.