Moved map out of tab control and add checks for lat-lon
The leaflet map would not render properly in a bootstrap tab control.
Re-designed the form so that the map was always visible.  It appears to
render properly now.  Added some functions to javascript to ensure that
lat-lon fields are constrained to values that make sense. For example minutes and
seconds must be positive and be between 0 and 60.