Added js to switch between metric and imperial units
Added additional inputs and associated java script to toggle between
metric and imperial units for total length, fork length, and round
weight.  When the values in these elements change, their counter part is
updated.  Also cleaned up some of the code associated with the leaflet
markers and maps.  Instead of destroying and recreating layers, we just
update the existing markers.  A separate marker is still required for
the popup map.