Added Add Another Tag button to new tag recovery
Modified the tag recovery detail view to include a button to add another
tag recovery event.  Makes it easier for users to add multiple tags
associated with same recovery report.  Added tests to verify that the
button is included in new tag recovery events, but not when existing
recover events are edited.