Final additions to implement report-a-tag workflow
This commit includes a number of updates to report-a-tag workflow.  The
urls have been grouped together and named using a standard prefix, tests
have been written to ensure that all of the urls are valid and that the
content rendered contains instructional messages when accessed through
the report-a-tag urls (and does contain a message otherwise). All of the
previously passing tests continue to pass (229/231).