Added report_count() and tag_count() methods to JoePublic
Added methods to JoePublic model to return the number of reports and
total number of tags returned by that individual.  Added tests to verify
that methods work as expected.  Updated the angler list to include the
number of reports and tags returned by each person.  Also added alert
box around message stating a match could not be found. If a match cannot
be found in report-a-tag process, next steps are provided.