Added action to recovery form view context
Added a variable 'action' to context of views that render recovery form
that indicates whether or not the form is associated with a new tag
recovery event or editing an existing one. The variable is used to
control whether imperial (new) or metric(existing) controls are
presented first.  Most of the code is written in javascript and cannot
be test using standard client but does appear to work as expected in my
current browser. Thank you LRC.