Default fate to K and updates to form html javascript
Modified the fate field of the recovery form to 'K' - most of our tagged
fish are walleye and most of them are harvested.  Released fish tend to
be the exception.  Also update the javascript to return a blank cell if
the conversion between metric and imperial returns NaN.  Still can't
thoroughly test javascript.  Additionally, switched FLEN and TLEN
elements on recovery form - most fish have a reported TLEN and no
FLEN. It makes the most sense to have the more common field first.