Commit of unstaged changes: preparing for rest-api
Most of the changes in this commit are associated with functional
tests.  Commit also includes a utility script to get encounter
information from ms access databases.  This is the version of the
application that is running on production.  this commit stages all
existing changes in preparation for a new branch which will use the
django-restframework to provide json data to map widgets.