Added if statement for marks and linebreaks for popup text
Added if statement so every template that creates leaflet markers and
tries to put them on a map. The if statement skips tagging or recovery
events that don't have coordinates. We aslo modified the recovery get_comments
function to include html linebreaks so that line breaks render in the
popups and recovery detail page.