Improved query performance - select related and Count
Updated the queries for tagging events, recaptures, and angler returns
to include select related clauses where ever possible to minimize the
number of queries run on each view.  Also included tagging events
where tagstat is either A or A2.  Updated the tags applied and
recovered in years and projectects to report the number of distinct
tagids.  Updated on test associated with popup text. All tests
continue to pass.