Fixed a typo and added distinct to query for recovered tags
There was a typo in the note for application events on the template
for tags recovered in a project that was causing confusion. The
template has now been updated.  Additional a DISTINCT clause was added
to the query that retrieves application events. If a tag was
recaptured more than once in a project - an application event was
being returned for each one. There should only be one application
event for each tag.