Updated utility script to append by year
This commit includes changes to the main utility script that appends
tag encounters. The script and the associated database have been
updated to append data in chunks representing a single calendar year.
This prevents us from having to rebuild the entire database when we
need to append new data, just data from the last year or two can be
updated.  The whole database can be rebuild if first year and last
year left unspecified.