Updates to get tests passing again.
This commit is mostly just a number of small changes to get the test
passing without issuing warnings. There were some urls in templates
that did not have the namespace.  This commit also includes a small
number of changes to the markdown letter to get the functionality
working again, but that functionality seems to be out of scope given
the non-standard approach to writing response letters.