Implemented most of the Follow work-flow including some tests
this commit is a big one - it includes a number of changes to the
views, forms, urls, tests and templates to accommodate the new report
follow-up work-flow that includes an intermediate step.  All of the
existing tests now pass. More tests need to be written to ensure that
report follow up work flow is being implemented properly. Next steps
include locking down all views with forms, added created and last
modified fields to report elements, and adding current status field to
the report model so that we can avoid many of the queries that are
currently being executed when some templates are rendered.