Added more report followup work flow tests and added next to form
Added a next parameter  to the work flow form and all of the views
that call it so that the user is returned to their origianl locaion
after the form is submitted or cancelled. Also added more tests to
ensure that the form behaves as expected, and that the available
choices dynamcally reflect the current status of the report.