Changed report.follow_up_status from method to field
Added a field to the report model to hold the current data status of
the report.  It use to be a method that would return the 'highest'
status from the report follow up table, but it was producing too many
queries when any template rendering the report status was called.
Changes to the report follow up status field and report status table
are wrapped in a transaction to ensure that they remain in sync.  All
of the existing tests were modified to reflect this new database
structure and continue to pass.