Removed code associated with the recovery letter creation
Removed the views, urls, and template files that were originally used
to create the recovery letters for angler reports. this functionality
is largely outside the scope of TFAT.  Also removed a little commented
code that was left over from creating the report.follow_up_status
field.  All of the existing tests continue to pass.