Added Lake filter to most views and a button bar widget to templates
Most of the views that present list of things or collections of things
now accept an optional lake parameter to filter the returned values.
A button bar was created and in included in most templates that will
allow the user to switch between or combine lakes easily. Additionally
a function was added to include the extents for the leaflet map so
that it could be set appropriately to remain in sync with the
selected filter.  All the existing tests pass, but the new filtering
functionality does not have associated tests yet.