Added TaggedSpecies model that inherits from common.species
The common species model is likely to have all species in the species
look-up table, most of which have never been, or ever will be
tagged. The inherited model TaggedSpecies allows us to reduce the
number species available in the TFAT application.  An admin interface
has been included to toggle the tagged status of species to have them
included or excluded as needed.  Tests were written
to ensure that the string representation methods and model managers
work as expected for the new TaggedSpecies model.  All existing tests
continue to pass.