Updated recovery form to get grid data from common api
The recovery form use to load a static file of grid coordinates to
verify and plot the 5-minute grid entered into the form. The form has
now been updated to call the common api to get a list of valid grids
and their centroids and to find the 5-minute grid corresponding to the
provided coordinates.  All existing tests continue to pass.