Added api endpoint urls and views for recovery letters
This commit includes several new urls and views that emulate the old
postgres sql queries that were used to create the tag recovery letters
and tag recovery and encounter records used in the walleye management
plan. These are not true api endpoints (they do not use django
restframework), but work now. They will serve as a good starting point
when we implement a real api for TFAT.  This commit also include a
skeleton of 11 failing tests associated with these endpoints. These
tests will be addressed in the next commit.