Added views, serializers and tests for angler api
This commit includes views, urls, serializers, and associated tests for
angler api endpoints. The tests verify that the readonly endpoints are
accessible and return what is expected. They are also ensure that the
crud endpoints are only accessible to authenticated users and preform as
expected.  All tests continue to pass.