Upgraded to Python 3.9 and Django==3.2
This commit updates tfat to Django==3.2.12 as well as all of the
associated dependencies. The single biggest change associated with the
update was replacing url() with path() and re_path() functions.  It
also removes several old migration files that has already been
squashed and updates a small number of tests to get them passing
again. All existing tests continue to pass, but the leaflet maps to
not appear to be rendering correctly.